Added cache.json, refactored notes.js
CACHE:
--Added a cache.json file and added helper functions to notes.js to handle errors.

Rationale is to move error checking to the front of each method to avoid
running queries on notes.json unnecessarily. In addition, the helper functions
fetchCache and queryTitle are more reusable and readable than prior code.

REFACTOR:
--renamed 'delete' method to 'remove' to avoid conflict with reserved words
and to better conform to CRUD conventions.
--moved CRUD methods into one constructor function to simplify scope and to
make it clearer what is being exported.